little tweaks & improvements
